move nonstandard into werift/nonstandard to minimize supply chain exposure and dependencies
