+++ packages/rtp/tests/processor/ntpTime.test.ts
@@ -1,12 +1,5 @@
-import {
-  int,
-  Max32Uint,
-  NtpTimeBase,
-  RtcpSenderInfo,
-  RtcpSrPacket,
-  RtpHeader,
-  RtpPacket,
-} from "../../src";
+import { RtcpSenderInfo, RtcpSrPacket, RtpHeader, RtpPacket } from "../../src";
+import { Max32Uint, NtpTimeBase } from "../../src/processor";
 
 describe("ntpTime", () => {
   it("rollover", () => {
